Move the keras dataset utilities to utils directory
PiperOrigin-RevId: 435400313
