+++ keras/preprocessing/BUILD
@@ -20,7 +20,6 @@ py_library(
         ":image",
         ":sequence",
         ":text",
-        ":timeseries",
         "//keras/utils:all_utils",
     ],
 )
@@ -28,9 +27,7 @@ py_library(
 py_library(
     name = "image",
     srcs = [
-        "dataset_utils.py",
         "image.py",
-        "image_dataset.py",
     ],
     srcs_version = "PY3",
     deps = [
@@ -57,23 +54,9 @@ py_library(
 )
 
 py_library(
-    name = "timeseries",
-    srcs = [
-        "timeseries.py",
-    ],
-    srcs_version = "PY3",
-    deps = [
-        "//:expect_numpy_installed",
-        "//:expect_tensorflow_installed",
-    ],
-)
-
-py_library(
     name = "text",
     srcs = [
-        "dataset_utils.py",
         "text.py",
-        "text_dataset.py",
     ],
     srcs_version = "PY3",
     deps = ["//:expect_tensorflow_installed"],
@@ -87,20 +70,6 @@ tf_py_test(
     tags = [
         "no_oss",  # TODO(scottzhu): Fix for multiple export issue.
     ],
-    deps = [
-        ":image",
-        "//:expect_numpy_installed",
-        "//:expect_tensorflow_installed",
-        "//keras",
-        "//keras/testing_infra:test_combinations",
-    ],
-)
-
-tf_py_test(
-    name = "image_dataset_test",
-    size = "small",
-    srcs = ["image_dataset_test.py"],
-    python_version = "PY3",
     deps = [
         ":image",
         "//:expect_numpy_installed",
@@ -133,32 +102,5 @@ tf_py_test(
         "//:expect_numpy_installed",
         "//:expect_tensorflow_installed",
         "//keras/testing_infra:test_combinations",
-    ],
-)
-
-tf_py_test(
-    name = "text_dataset_test",
-    size = "small",
-    srcs = ["text_dataset_test.py"],
-    python_version = "PY3",
-    deps = [
-        ":text",
-        "//:expect_tensorflow_installed",
-        "//keras",
-        "//keras/testing_infra:test_combinations",
-    ],
-)
-
-tf_py_test(
-    name = "timeseries_test",
-    size = "small",
-    srcs = ["timeseries_test.py"],
-    python_version = "PY3",
-    deps = [
-        ":timeseries",
-        "//:expect_numpy_installed",
-        "//:expect_tensorflow_installed",
-        "//keras/testing_infra:test_combinations",
-        "//keras/testing_infra:test_utils",
     ],
 )
