+++ keras/preprocessing/__init__.py
@@ -31,5 +31,4 @@ from keras import backend
 from keras.preprocessing import image
 from keras.preprocessing import sequence
 from keras.preprocessing import text
-from keras.preprocessing import timeseries
 from keras.utils import all_utils as utils
