+++ keras/utils/BUILD
@@ -33,8 +33,11 @@ py_library(
     deps = [
         ":data_utils",
         ":generic_utils",
+        ":image_dataset",
         ":layer_utils",
         ":np_utils",
+        ":text_dataset",
+        ":timeseries_dataset",
         ":vis_utils",
     ],
 )
@@ -237,6 +240,46 @@ py_library(
     ],
     srcs_version = "PY3",
     deps = [
+        "//:expect_tensorflow_installed",
+    ],
+)
+
+py_library(
+    name = "image_dataset",
+    srcs = [
+        "dataset_utils.py",
+        "image_dataset.py",
+    ],
+    srcs_version = "PY3",
+    deps = [
+        "//:expect_numpy_installed",
+        "//:expect_tensorflow_installed",
+        "//keras/layers/preprocessing:image_preprocessing",
+        "//keras/preprocessing:image",
+    ],
+)
+
+py_library(
+    name = "text_dataset",
+    srcs = [
+        "dataset_utils.py",
+        "text_dataset.py",
+    ],
+    srcs_version = "PY3",
+    deps = [
+        "//:expect_numpy_installed",
+        "//:expect_tensorflow_installed",
+    ],
+)
+
+py_library(
+    name = "timeseries_dataset",
+    srcs = [
+        "timeseries_dataset.py",
+    ],
+    srcs_version = "PY3",
+    deps = [
+        "//:expect_numpy_installed",
         "//:expect_tensorflow_installed",
     ],
 )
@@ -450,5 +493,46 @@ tf_py_test(
     deps = [
         "//:expect_tensorflow_installed",
         "//keras",
+    ],
+)
+
+tf_py_test(
+    name = "image_dataset_test",
+    size = "small",
+    srcs = ["image_dataset_test.py"],
+    python_version = "PY3",
+    deps = [
+        ":image_dataset",
+        "//:expect_numpy_installed",
+        "//:expect_tensorflow_installed",
+        "//keras",
+        "//keras/testing_infra:test_combinations",
+    ],
+)
+
+tf_py_test(
+    name = "text_dataset_test",
+    size = "small",
+    srcs = ["text_dataset_test.py"],
+    python_version = "PY3",
+    deps = [
+        ":text_dataset",
+        "//:expect_tensorflow_installed",
+        "//keras",
+        "//keras/testing_infra:test_combinations",
+    ],
+)
+
+tf_py_test(
+    name = "timeseries_dataset_test",
+    size = "small",
+    srcs = ["timeseries_dataset_test.py"],
+    python_version = "PY3",
+    deps = [
+        ":timeseries_dataset",
+        "//:expect_numpy_installed",
+        "//:expect_tensorflow_installed",
+        "//keras/testing_infra:test_combinations",
+        "//keras/testing_infra:test_utils",
     ],
 )
