+++ keras/utils/timeseries_dataset_test.py
@@ -12,13 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Tests for timeseries."""
+"""Tests for timeseries_dataset."""
 
 import tensorflow.compat.v2 as tf
 
 import numpy as np
 from keras.testing_infra import test_utils
-from keras.preprocessing import timeseries
+from keras.utils import timeseries_dataset
 
 
 @test_utils.run_v2_only
@@ -28,7 +28,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
     # Test ordering, targets, sequence length, batch size
     data = np.arange(100)
     targets = data * 2
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, targets, sequence_length=9, batch_size=5)
     # Expect 19 batches
     for i, batch in enumerate(dataset):
@@ -50,7 +50,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
     data = np.arange(10)
     offset = 3
     targets = data[offset:]
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, targets, sequence_length=offset, batch_size=1)
     i = 0
     for batch in dataset:
@@ -65,7 +65,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
 
   def test_no_targets(self):
     data = np.arange(50)
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, None, sequence_length=10, batch_size=5)
     # Expect 9 batches
     i = None
@@ -83,7 +83,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
     # Test cross-epoch random order and seed determinism
     data = np.arange(10)
     targets = data * 2
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, targets, sequence_length=5, batch_size=1, shuffle=True, seed=123)
     first_seq = None
     for x, y in dataset.take(1):
@@ -94,7 +94,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
     for x, _ in dataset.take(1):
       self.assertNotAllClose(x, first_seq)
     # Check determism with same seed
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, targets, sequence_length=5, batch_size=1, shuffle=True, seed=123)
     for x, _ in dataset.take(1):
       self.assertAllClose(x, first_seq)
@@ -102,7 +102,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
   def test_sampling_rate(self):
     data = np.arange(100)
     targets = data * 2
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, targets, sequence_length=9, batch_size=5, sampling_rate=2)
     for i, batch in enumerate(dataset):
       self.assertLen(batch, 2)
@@ -123,7 +123,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
   def test_sequence_stride(self):
     data = np.arange(100)
     targets = data * 2
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, targets, sequence_length=9, batch_size=5, sequence_stride=3)
     for i, batch in enumerate(dataset):
       self.assertLen(batch, 2)
@@ -144,7 +144,7 @@ class TimeseriesDatasetTest(tf.test.TestCase):
 
   def test_start_and_end_index(self):
     data = np.arange(100)
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, None,
         sequence_length=9, batch_size=5, sequence_stride=3, sampling_rate=2,
         start_index=10, end_index=90)
@@ -155,31 +155,31 @@ class TimeseriesDatasetTest(tf.test.TestCase):
   def test_errors(self):
     # bad start index
     with self.assertRaisesRegex(ValueError, '`start_index` must be '):
-      _ = timeseries.timeseries_dataset_from_array(
+      _ = timeseries_dataset.timeseries_dataset_from_array(
           np.arange(10), None, 3, start_index=-1)
     with self.assertRaisesRegex(ValueError, '`start_index` must be '):
-      _ = timeseries.timeseries_dataset_from_array(
+      _ = timeseries_dataset.timeseries_dataset_from_array(
           np.arange(10), None, 3, start_index=11)
     # bad end index
     with self.assertRaisesRegex(ValueError, '`end_index` must be '):
-      _ = timeseries.timeseries_dataset_from_array(
+      _ = timeseries_dataset.timeseries_dataset_from_array(
           np.arange(10), None, 3, end_index=-1)
     with self.assertRaisesRegex(ValueError, '`end_index` must be '):
-      _ = timeseries.timeseries_dataset_from_array(
+      _ = timeseries_dataset.timeseries_dataset_from_array(
           np.arange(10), None, 3, end_index=11)
     # bad sampling_rate
     with self.assertRaisesRegex(ValueError, '`sampling_rate` must be '):
-      _ = timeseries.timeseries_dataset_from_array(
+      _ = timeseries_dataset.timeseries_dataset_from_array(
           np.arange(10), None, 3, sampling_rate=0)
     # bad sequence stride
     with self.assertRaisesRegex(ValueError, '`sequence_stride` must be '):
-      _ = timeseries.timeseries_dataset_from_array(
+      _ = timeseries_dataset.timeseries_dataset_from_array(
           np.arange(10), None, 3, sequence_stride=0)
 
   def test_not_batched(self):
     data = np.arange(100)
 
-    dataset = timeseries.timeseries_dataset_from_array(
+    dataset = timeseries_dataset.timeseries_dataset_from_array(
         data, None, sequence_length=9, batch_size=None, shuffle=True)
     sample = next(iter(dataset))
     self.assertEqual(len(sample.shape), 1)
